Upload fixed file
Create project ANKETA. Ask the user:
surname, name, second name. SEPARATE (by the prompt operator)
age in years (by the prompt operator)
gender (with a confirm operator, for example, "is your gender male?")
and display the user's profile with the alert operator, for example:

your name: Ivanov Ivan Ivanovich
your age in years: 20
your age in days: 7300
in 5 years you will be: 25
your gender: male
you are retired: no
There should be control over the correctness of the data entered by the user (for example, the surname should not be empty, the age should be a correct number, etc.).
The alert statement must be used exactly once in your code.