Upload fixed file
Develop a "pure" function treeSum that receives an array whose elements can be numbers or again arrays, and so on to any level. 
The function should calculate and return the sum of all numeric elements of the array from all levels. 
When writing a function, do not describe any functions nested in it.