Add file of additional task
Write a clean function that checks that the phrase passed to it is a palindrome.
Do not use arrays in the decision.
When checking, the following should be ignored:
- case of letters;
- spaces;
- punctuation marks;
Ask the user for a string. Display the message "this is a palindrome" or "this is not a palindrome" via alert.