Add file for task 7 Drinks
Develop a HashStorageFunc class in a functional style (i.e. a constructor function) to store arbitrary key-value pairs in a hash.
The key can be any string; the value can be of any type, including vector (hash, array, etc.) The class must have the following interface (i.e., have the following public methods):
addValue (key, value) - stores the specified value under the specified key;
getValue (key) - returns the value for the specified key or undefined;
deleteValue (key) - deletes the value with the specified key, returns true if the value was deleted and false if there was no such value in the storage;
getKeys () - returns an array consisting of some keys. The class must be clean (must not use any global variables, must not "dirty the screen").
The class must be generic, i.e. not depend on the structure of the stored data, nor on the method of their subsequent use (including should not contain any links
to the DOM, because can be used without a web page at all).