Upload files for task#11 Drag_n_Drop
Implement in JavaScript to drag and drop any (but not small) images with the mouse on a web page. Handle at least mousedown, mousemove, mouseup events. Images should be “dragged” by the mouse over any point (that is, both when “picking up” and “releasing” the image should not move, it should only move when the mouse is moved with the left button pressed, exactly as far as the mouse is displaced).
The code should not depend on the number of images (i.e. the code should find all the images in the specified div-container by itself).
Pictures should NOT initially be positioned (the position style property should not be set). When you start dragging a picture, the rest of the pictures should not move. The picture you started dragging should appear higher (closer to the eyes) than the rest. While dragging, change the cursor shape to a suitable one.